Moving repository search annotations to a separate submodule
