+++ src/services/repositories.py
@@ -1,48 +1,5 @@
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
 from typing import Optional, Iterable, Union, Callable
-
-
-class SearchAnnotation:
-    pass
-
-
-@dataclass(frozen=True)
-class ValueAnnotation(SearchAnnotation):
-    value: any
-
-
-@dataclass(frozen=True)
-class Greater(ValueAnnotation):
-    pass
-
-
-@dataclass(frozen=True)
-class Lesser(ValueAnnotation):
-    pass
-
-
-@dataclass(frozen=True)
-class Equal(ValueAnnotation):
-    pass
-
-
-@dataclass(frozen=True)
-class Not(ValueAnnotation):
-    pass
-
-
-class GroupingAnnotation(SearchAnnotation):
-    def __init__(self, *annotations: SearchAnnotation):
-        self.annotations = annotations
-
-
-class And(GroupingAnnotation):
-    pass
-
-
-class Or(GroupingAnnotation):
-    pass
 
 
 class IRepository(ABC):
